Keep reference to "original" CL
As that one is needed for dynamic construction of some
inner bits, while the while listed should be used
during runtime.

This change basically modifies only the two "createDynamically"
method, that does Class.forName calls.

This affect XStream in Nexus, as then white listing
is enabled and Nexus fails to boot (as shown by ITs)